Add current memory consumption counter

+++ assembler/src/include/perfcounter.hpp
@@ -53,13 +53,12 @@ inline std::string human_readable_time(double time_in_sec)
     return str(boost::format("%3d:%02d:%02d.%03d") % hours % mins % sec % msec);
 }
 
-inline std::string human_readable_memory(unsigned max_rss)
-{
+inline std::string human_readable_memory(size_t max_rss) {
 	assert(max_rss > 0);
 
 	if (max_rss < 1024 * 1024) {
 		return str(boost::format("%4dM") % (max_rss / 1024));
-    } else {
+  } else {
     return str(boost::format("%4dG") % (max_rss / (1024 * 1024)));
 	}
 }
